feat(devices): add device type selection with tabbed interface
Add integrated device type selection directly in the devices page, allowing users to choose between local and cloud devices using a tabbed interface.

Changes:
- Add DeviceSetupGuide component with tabbed interface (local/cloud)
- Refactor devices page to use new setup guide
- Update DevicesPageHeader to simplify "Add Device" button
- Remove LocalExecutorGuide's internal border and padding
- Add translations for device type selection
- Extract device handlers to custom hook (useDeviceHandlers)
- Add device status utility functions
