+++ frontend/src/features/devices/hooks/index.ts
@@ -1,0 +1,9 @@
+// SPDX-FileCopyrightText: 2025 Weibo, Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+/**
+ * Device hooks barrel export.
+ */
+
+export * from './useDeviceHandlers'
