+++ frontend/src/i18n/locales/zh-CN/devices.json
@@ -105,5 +105,16 @@
   "local_devices_section": "本地设备",
   "cloud_devices_section": "云端设备",
   "no_local_devices": "暂无本地设备连接。请使用上方指南添加设备。",
-  "cloud_devices_coming_soon": "云端设备支持即将推出。"
+  "cloud_devices_coming_soon": "云端设备支持即将推出。",
+  "select_device_type": "选择设备类型",
+  "select_device_type_description": "选择您想要添加的设备类型",
+  "local_device": "本地设备",
+  "local_device_description": "在您自己的机器上运行 Executor，完全掌控和隐私保护",
+  "cloud_device": "云端设备",
+  "cloud_device_description": "使用 Wegent 托管的云端基础设施（无需安装配置）",
+  "recommended": "推荐",
+  "continue": "继续",
+  "cloud_coming_soon_title": "云端设备即将推出",
+  "cloud_coming_soon_description": "云端设备支持正在积极开发中，敬请期待！",
+  "got_it": "知道了"
 }
